requirements.txt
absl-py==1.0.0
ale-py==0.7.1
asl-pepper-2d-sim==0.0.5
astor==0.8.1
astroid==2.9.3
atari-py==0.2.9
backcall==0.2.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bleach==1.5.0
cached-property==1.5.2
cachetools==4.2.4
catkin-pkg==0.4.24
certifi==2016.9.26
charset-normalizer==2.0.12
cloudpickle==1.6.0
commonmark==0.9.1
cupy==9.6.0
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docutils==0.18.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
fastrlock==0.8.2
flake8==4.0.1
future==0.18.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.19.0
h5py==3.1.0
html5lib==0.9999999
idna==3.4
imageio==2.15.0
importlib-metadata==4.2.0
importlib-resources==5.4.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912934572/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython==7.16.3
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.17.2
joblib==1.1.0
jupyter-client==5.3.4
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
mock==4.0.3
nav-msgs==1.13.0.post3
navrep==0.0.14
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.1.0
parso==0.7.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.5
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyglet==1.5.22
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyIAN==0.0.3
PyKDL==1.4.0
pylidar2d==0.1
pylint==2.12.2
pymap2d==0.1.15
pyniel==0.2
pyparsing==3.0.7
pyrangelibc-danieldugas==0.0.2
pyrvo2-danieldugas==0.0.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
PyYAML==6.0
pyzmq==19.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.0.0
rospkg==1.4.0
rsa==4.9
scipy==1.5.4
simplegeneric==0.8.1
sip==6.5.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
snakeviz==2.1.1
soadrl-crowdnav==0.0.3
stable-baselines==2.10.0
stable-baselines3==1.3.0
std-srvs==1.11.3.post1
strictfire==0.4.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==1.13.2
tensorflow-estimator==1.13.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toml==0.10.2
torch==1.8.2+cu111
torchaudio==0.8.2
torchvision==0.9.2+cu111
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm==4.63.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1631041982274/work
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.16
visualization-msgs==1.13.0.post4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1696255154857/work
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.6.0
